# cache: Allow sharing compression variants among refs