--- a/cache/remote.go
+++ b/cache/remote.go
@@ -12,6 +12,7 @@ import (
 	"github.com/moby/buildkit/cache/config"
 	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/solver"
+	"github.com/moby/buildkit/util/bklog"
 	"github.com/moby/buildkit/util/compression"
 	"github.com/moby/buildkit/util/contentutil"
 	"github.com/moby/buildkit/util/leaseutil"
@@ -53,7 +54,7 @@ func (sr *immutableRef) GetRemotes(ctx context.Context, createIfNeeded bool, ref
 	// compression with all combination of copmressions
 	res := []*solver.Remote{remote}
 	topmost, parentChain := remote.Descriptors[len(remote.Descriptors)-1], remote.Descriptors[:len(remote.Descriptors)-1]
-	vDesc, err := getCompressionVariantBlob(ctx, sr.cm.ContentStore, topmost.Digest, refCfg.Compression.Type)
+	vDesc, err := getBlobWithCompression(ctx, sr.cm.ContentStore, topmost, refCfg.Compression.Type)
 	if err != nil {
 		return res, nil // compression variant doesn't exist. return the main blob only.
 	}
@@ -108,16 +109,16 @@ func getAvailableBlobs(ctx context.Context, cs content.Store, chain *solver.Remo
 	if err != nil {
 		return nil, err
 	}
-	compressions, err := getCompressionVariants(ctx, cs, target.Digest)
-	if err != nil {
-		return nil, err
+	var descs []ocispecs.Descriptor
+	if err := walkBlob(ctx, cs, target, func(desc ocispecs.Descriptor) bool {
+		descs = append(descs, desc)
+		return true
+	}); err != nil {
+		bklog.G(ctx).WithError(err).Warn("failed to walk variant blob") // is not a critical error at this moment.
 	}
 	var res []*solver.Remote
-	for _, c := range compressions {
-		desc, err := getCompressionVariantBlob(ctx, cs, target.Digest, c)
-		if err != nil {
-			return nil, err
-		}
+	for _, desc := range descs {
+		desc := desc
 		if len(parents) == 0 { // bottommost ref
 			res = append(res, &solver.Remote{
 				Descriptors: []ocispecs.Descriptor{desc},
@@ -217,9 +218,9 @@ func (sr *immutableRef) getRemote(ctx context.Context, createIfNeeded bool, refC
 			} else if needs {
 				// ensure the compression type.
 				// compressed blob must be created and stored in the content store.
-				blobDesc, err := ref.getCompressionBlob(ctx, refCfg.Compression.Type)
+				blobDesc, err := getBlobWithCompressionWithRetry(ctx, ref, refCfg.Compression, s)
 				if err != nil {
-					return nil, errors.Wrapf(err, "compression blob for %q not found", refCfg.Compression.Type)
+					return nil, errors.Wrapf(err, "failed to get compression blob %q", refCfg.Compression.Type)
 				}
 				newDesc := desc
 				newDesc.MediaType = blobDesc.MediaType
@@ -249,6 +250,16 @@ func (sr *immutableRef) getRemote(ctx context.Context, createIfNeeded bool, refC
 		})
 	}
 	return remote, nil
+}
+
+func getBlobWithCompressionWithRetry(ctx context.Context, ref *immutableRef, comp compression.Config, s session.Group) (ocispecs.Descriptor, error) {
+	if blobDesc, err := ref.getBlobWithCompression(ctx, comp.Type); err == nil {
+		return blobDesc, nil
+	}
+	if err := ensureCompression(ctx, ref, comp, s); err != nil {
+		return ocispecs.Descriptor{}, errors.Wrapf(err, "failed to get and ensure compression type of %q", comp.Type)
+	}
+	return ref.getBlobWithCompression(ctx, comp.Type)
 }
 
 type lazyMultiProvider struct {
@@ -300,6 +311,11 @@ func (p lazyRefProvider) Unlazy(ctx context.Context) error {
 		} else if !isLazy {
 			return nil, nil
 		}
+		defer func() {
+			if rerr == nil {
+				rerr = p.ref.linkBlob(ctx, p.desc)
+			}
+		}()
 
 		if p.dh == nil {
 			// shouldn't happen, if you have a lazy immutable ref it already should be validated
@@ -334,14 +350,6 @@ func (p lazyRefProvider) Unlazy(ctx context.Context) error {
 			}
 		}
 
-		compressionType := compression.FromMediaType(p.desc.MediaType)
-		if compressionType == compression.UnknownCompression {
-			return nil, errors.Errorf("unhandled layer media type: %q", p.desc.MediaType)
-		}
-
-		if err := p.ref.addCompressionBlob(ctx, p.desc, compressionType); err != nil {
-			return nil, err
-		}
 		return nil, nil
 	})
 	return err