… cache/manager_test.go