Fixes: #2346

Currently, a compression variant of a ref is not visible from another. This leads to duplicated conversion as mentioned in #2346.
This commit allows refs sharing compression variants with each other.

When a ref converts a blob (compression blob A) to another compression (compression blob B), Blob A and B are linked together using GC label to make sure that all linked variants are released together (i.e. they aren't released even when refs lease only some of them). (Maybe in the future we can switch to the lease-based approach as discussed in  #2347. But how can we manage the ref counter of the lease itself among refs in a simple way?)

When a ref exports a compression variant, it first walks all linked compression variant blobs and checks if there is a blob that can be reused without conversion. Conversion happens only when no blob can be reused.

## latency of exporing layers

This runs two builds with two compression types (gzip and zstd). Both of the builds share the same lower layers.

A:

```dockerfile
FROM ${REGISTRY}/${IMG}
RUN echo hello > /hello
```

B: 

```dockerfile
FROM ${REGISTRY}/${IMG}
RUN echo hi > /hi
```

Before the measurement, zstd-compressed B is stored to the registry.

- Steps: https://github.com/ktock/buildkit/blob/5b9997d2b45ce57ed8ad07969840069e1292a388/hack/blob/measure.sh#L51-L71
  - First, we build A as gzip then convert it to zstd.
    - Here, gzip and zstd variants of `${REGISTRY}/${IMG}` are stored to the content store.
  - Next, we pull zstd-compressed B (contains zstd-version of `${REGISTRY}/${IMG}` in the lower layers) from the registry then convert it to gzip.
    - Here, gzip-version of `${REGISTRY}/${IMG}` has already been cached during building A, so this build should reuse these blobs without duplicated conversions.
    - This PR-version of BuildKit successfully reuses it but master version of BuildKit doesn't.

The following shows the time to export layers during building gzip-version of B at the final step in the above.
This PR exports layers without conversion so it's faster than the master version.

- measured on Github Actions ubuntu runner:
  - PR: https://github.com/ktock/buildkit/actions/runs/1805493851
  - master: https://github.com/ktock/buildkit/actions/runs/1805496498

### OCI worker

|`${IMG}`|master|PR|
|---|---|---|
|ubuntu:20.04|4.1s|0s|
|tomcat:10.0.0-jdk15-openjdk-buster|23.3s|0s|

### containerd worker

|`${IMG}`|master|PR|
|---|---|---|
|ubuntu:20.04|25.5s|0s|
|tomcat:10.0.0-jdk15-openjdk-buster|190.2s|0.1s|

master version of containerd worker's performance is too worse than OCI worker.
This is because of the lack of bufio writers for the content store.
This will be fixed by #2601.